better implementation of console.ToWriter (#2697)
It's so much better that it's been renamed to `console.ToWriteCloser`.

The problem was that writes from exec.Cmd are buffered, so you end up with weird extra or missing newlines in the log. This change instead pipes the output to a buffered reader, and then reads lines on it until it gets EOF. We return a write-closer instead of a writer, so we can (1) close the underlying writer and (2) make sure all the data has been written out from the reader before closing.

This makes `RunCmdInConsole` an interesting enough function to move into the `exec` package.

fbshipit-source-id: 6a7f43a