[api/client] add more identifying request headers (#2878)
I'm not totally happy with how this change turned out, but I tried a few other ways that were worse.

The goal was to include the following three headers in all requests from the api client:

```go
// ClientApp is the name of the application calling the API
ClientApp = "X-Tbn-Api-Client-App"

// ClientType indicates which kind of API client is making a request (go http
// client vs js web client etc.)
ClientType = "X-Tbn-Api-Client-Type"

// ClientVersion indicates the version of the API client
ClientVersion = "X-Tbn-Api-Client-Version"
```

This way, we know the library used, the version of the library, and the particular caller (eg tbnctl, tbnproxy, etc). I'm not doing anything with these headers, though I'd like to include them in request logs for the api and stats servers.

ClientType and ClientVersion are hidden underneath the api client code, so there wasn't much work there. ClientApp was a pain, because it needs to be threaded down.

In implementing this, the choices were:

1. change various `NewFromFlags` calls to include a client app
2. change various `FromFlags.Make()` calls to include a client app

I did (2) first, and it was much worse because `Make()` is heavily mocked and tested, and this really isn't something you need to _configure_ from flags, it's something known at compile time, so it made more sense to me to have it be part of FromFlags instantiation rather than the Make call.

My next choice was whether to use `Options` in `FromFlags` that already had them, vs making it a new required field and changing the API. In all but one case, I chose the latter, because it's not really optional, it's required, and the `Options` pattern doesn't handle this well (you can make it blow up at Validation time, but again, this should all be well-known ahead of time).

The one execption was the threading of the StatsClientFromFlags into `logparser/forwarder/turbineForwarderFromFlags`; It was certainly possible to thread this all the way through the NewFromFlags stack, but it was a pretty big onion, and since it was only relevant when using a stats client, it didn't seem worth it. If you use the api forwarder without setting the client explicitly, you'll get a default client app, and that seems fine to me.

It might be easiest to read this from the bottom up, so you can see how calling code is affected.

fbshipit-source-id: da7e492