[tbnctl] add support for domain aliases in init-zone (fixes #4106) (#4117)
The point of this change is to let us make "localhost" an alias for "all-in-one-demo" so that we don't need to add "all-in-one-demo" to `/etc/hosts`.

To do it, add a `--domains` flag to `tbnctl init-zone`, which lets you specify a list of aliases for a domain.

Also switch a bunch of Error() logging to Debug() logging, because it's not an error.

fbshipit-source-id: e9b8d2c