[tbnctl] "fix" error handling (fixes #3449) (#4118)
The fix is in quotes because I actually get _less_ clever, not more. The previous code would attempt to encode errors to json, which works for the class of errors that come straight from the API, but not for vanilla errors created with `errors.New` or `fmt.Errorf`. These can still be marshaled to JSON, but they all look like `{}` because they have no public json fields.

The other problem with the old approach is that API errors do not return in non-zero exit codes, because they are marshaled just as a non-error object would be.

The other other problem is that the contract of our CLI apps is to have subcommand errors be prefixed by the name of the subcommand; there's no expectation of JSON-compliant error output, and it's a huge pain to make all the errors work that way.

SOOOOO, instead i just let errors be errors, and use `cmd.Error(err)` on the failure case, and call it a day.

update: JSON/YAML-encode api/http/error.Errors, let others be.

fbshipit-source-id: d75bf3a