add `--deep` flag to `tbnctl delete` (fixes #2607) (#4151)
wow this turned out to be a lot of code.

The idea here is to be able to do a cascading delete of certain objects:

Zone: everything
Domain: Routes, SharedRules (if orphaned), Modify Proxies to remove the Domain
Routes: SharedRules (if orphaned)
SharedRules: Routes
Proxy: n/a
Cluster: n/a
User: n/a

To do it, I made a `deleter` struct, which takes care of traversing objects graphs when necessary, reporting on what it plans to do, and logging as it proceeds.

In the GIF below, I add a zone, ponder deleting it, decide to delete the domain instead (so you can see the special treatment for the proxy), then delete the whole zone.

![deep_delete](https://user-images.githubusercontent.com/362808/32625747-f681b0fa-c541-11e7-9854-8d03f006a6a8.gif)

fbshipit-source-id: 89979f6