[tbnctl] zone import and export (fixes #2606) (#4288)
I've had this sitting in my tree for a long time, I think it's pretty much ready to go. It's not well unit-tested; It can be, but it's a huge pain in the ass. Instead, you can take my word for the fact that I exported and imported our zone a bunch of times, and it works great.

I wanted to merge this in now, because I'd like to get back to the task of building a new kube cluster, and this will be much easier with this in place. (Yes I can run it from my branch, but I think everyone should have the goodness)

change:

add a zoneObjects struct, which holds all the objects in a zone, and the maps needed for referential integrity
add importZone and exportZone functions, which do about what you'd expect
add corresponding commands

fbshipit-source-id: 80447c1